Service oriented design for better testability
