+++ app/services/application_service.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class ApplicationService
+  def self.call(*args, &block)
+    new(*args, &block).call
+  end
+end
