+++ app/spec/github_api_sample_responses/all_checks_successfully_completed.json
